+++ src/virtual-keyboard/utils.ts
@@ -781,6 +781,13 @@ const KEYCAP_SHORTCUTS: Record<string, Partial<VirtualKeyboardKeycap>> = {
     class: 'big-op hide-shift',
     label: '.',
   },
+  '[,]': {
+    variants: ',',
+    command: ['performWithFeedback', 'insertDecimalSeparator'],
+    shift: '.',
+    class: 'big-op hide-shift',
+    label: ',',
+  },
   '[+]': {
     variants: [{ latex: '\\sum_{#0}^{#0}', class: 'small' }, '\\oplus'],
     latex: '+',
@@ -986,6 +993,9 @@ export function normalizeKeycap(
   keycap: string | Partial<VirtualKeyboardKeycap>
 ): Partial<VirtualKeyboardKeycap> {
   if (typeof keycap === 'string') {
+    if (keycap === '[.]' && window.MathfieldElement.decimalSeparator === ',')
+      keycap = '[,]';
+
     if (!KEYCAP_SHORTCUTS[keycap]) return { latex: keycap };
     keycap = { label: keycap };
   }